+++ include/lwip/apps/snmp.h
@@ -83,7 +83,7 @@ void snmp_set_device_enterprise_oid(const struct snmp_obj_id* device_enterprise_
 const struct snmp_obj_id* snmp_get_device_enterprise_oid(void);
 
 void snmp_trap_dst_enable(u8_t dst_idx, u8_t enable);
-void snmp_trap_dst_ip_set(u8_t dst_idx, const ip_addr_t *dst);
+void snmp_trap_dst_ip_set(u8_t dst_idx, const struct net_in_addr *dst);
 
 /** Generic trap: cold start */
 #define SNMP_GENTRAP_COLDSTART 0