+++ include/lwip/ip.h
@@ -1,5 +1,0 @@
-/*
- * A stub for the ip.h header file.
- */
-
-#include "lwip/def.h"