+++ src/snmp_mib2_snmp.c
@@ -33,8 +33,8 @@
  *         Christiaan Simons <christiaan.simons@axon.tv>
  */
 
-#include "lwip/snmp.h"
 #include "lwip/apps/snmp.h"
+#include "lwip/def.h"
 #include "lwip/apps/snmp_core.h"
 #include "lwip/apps/snmp_mib2.h"
 #include "lwip/apps/snmp_scalar.h"